refactor: add payment result args and confirm payment args to args package
